zadanie2/flink-app/requirements.txt
apache-flink==1.18.1
kafka-python==2.0.2
confluent-kafka==2.2.0
clickhouse-driver==0.2.6
redis==5.0.1
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
pydantic==2.5.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
pyyaml==6.0.1
